+++ app/src/main/java/woowacourse/paint/view/CanvasView.kt
@@ -9,10 +9,10 @@ import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
 import woowacourse.paint.model.BrushType
-import woowacourse.paint.model.Line
+import woowacourse.paint.model.line.Line
 import woowacourse.paint.model.MyColor
-import woowacourse.paint.model.Rectangle
-import woowacourse.paint.model.RectangleVertex
+import woowacourse.paint.model.rectangle.Rectangle
+import woowacourse.paint.model.rectangle.RectangleVertex
 import woowacourse.paint.model.Sketch
 import woowacourse.paint.model.circle.Center
 import woowacourse.paint.model.circle.Circle
